Update glide_ide.py
Toolbox updated to include sectioned (topic) drop downs that will display the corresponding feature. Changed behavior of action when user clicks on a feature in a section to display a preview and added a confirm button. Added quick insert which will allows the user to click the feature and instantly paste to code editor (overriding the preview and confirm steps. Lasty, an undo button is added to undo one action (needs work).